Nessie-GC: Support Google Cloud Storage (GCS) (experimental)
Adding support for Google Cloud Storage as experimental. Unfortunately there's no way to test the integration in CI, because there's no emulator available.
