+++ gc/gc-iceberg-inttest/build.gradle.kts
@@ -28,6 +28,7 @@ dependencies {
   implementation(libs.hadoop.client)
   implementation(libs.iceberg.core)
   implementation(libs.iceberg.aws)
+  implementation(libs.iceberg.gcp)
 
   compileOnly(libs.errorprone.annotations)
   compileOnly(libs.microprofile.openapi)
@@ -77,6 +78,7 @@ dependencies {
 
   intTestRuntimeOnly(libs.iceberg.hive.metastore)
   intTestRuntimeOnly(libs.iceberg.aws)
+  intTestRuntimeOnly(libs.iceberg.gcp)
 
   intTestRuntimeOnly(libs.hadoop.client)
   intTestRuntimeOnly(libs.hadoop.aws)
@@ -89,6 +91,11 @@ dependencies {
   intTestRuntimeOnly("software.amazon.awssdk:dynamodb")
   intTestRuntimeOnly("software.amazon.awssdk:glue")
   intTestRuntimeOnly("software.amazon.awssdk:kms")
+
+  intTestImplementation(platform(libs.google.cloud.storage.bom))
+  intTestImplementation("com.google.cloud:google-cloud-storage")
+  intTestRuntimeOnly(libs.google.cloud.nio)
+  intTestRuntimeOnly(libs.google.cloud.gcs.connector)
 
   intTestCompileOnly(libs.immutables.builder)
   intTestCompileOnly(libs.immutables.value.annotations)
